Upd to laravel 11 and add improvement

+++ car-app/database/migrations/2023_11_01_142030_create_vehicles_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('vehicle_register_number');
             $table->text('other');
 
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id'); // Change to unsignedBigInteger for consistency
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
