+++ car-app/database/migrations/2023_11_01_143236_create_services_history_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('description');
             $table->string('cost');
 
-            $table->unsignedInteger('vehicle_id');
+            $table->unsignedBigInteger('vehicle_id');
             $table->foreign('vehicle_id')
                 ->references('id')
                 ->on('vehicles')
